+++ CHANGELOG.markdown
@@ -1,3 +1,4 @@
 # leaderboard 2.0.0 (in progress)
 
   * Change `addMember` to `rankMember`
+  * Added `deleteLeaderboard` and `deleteLeaderboardNamed`